+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## main
 
+* Create `IconButton` component.
+
+    *Manuel Puyol*
+
 ## 0.0.38
 
 * Extract `BaseButton` component.